+++ flutter_inappwebview_platform_interface/lib/src/types/client_cert_response.g.dart
@@ -16,9 +16,19 @@ class ClientCertResponse {
   String? androidKeyStoreType;
 
   ///The certificate password.
+  ///
+  ///**Officially Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
+  ///- MacOS
   String? certificatePassword;
 
   ///The file path of the certificate to use.
+  ///
+  ///**Officially Supported Platforms/Implementations**:
+  ///- Android native WebView
+  ///- iOS
+  ///- MacOS
   String certificatePath;
 
   ///An Android-specific property used by Java [KeyStore](https://developer.android.com/reference/java/security/KeyStore) class to get the instance.
@@ -26,14 +36,21 @@ class ClientCertResponse {
   ///**Officially Supported Platforms/Implementations**:
   ///- Android native WebView
   String? keyStoreType;
+
+  ///The index of the selected certificate.
+  ///
+  ///**Officially Supported Platforms/Implementations**:
+  ///- Windows
+  int selectedCertificate;
   ClientCertResponse(
-      {required this.certificatePath,
+      {this.certificatePath = "",
       this.certificatePassword = "",
       @Deprecated('Use keyStoreType instead')
       this.androidKeyStoreType = "PKCS12",
       this.keyStoreType = "PKCS12",
+      this.selectedCertificate = -1,
       this.action = ClientCertResponseAction.CANCEL}) {
-    if (this.action == ClientCertResponseAction.PROCEED)
+    if (this.action == ClientCertResponseAction.PROCEED && !Util.isWindows)
       assert(certificatePath.isNotEmpty);
     this.keyStoreType = this.keyStoreType ?? this.androidKeyStoreType;
     if (Util.isAndroid) assert(this.keyStoreType != null);
@@ -44,13 +61,17 @@ class ClientCertResponse {
     if (map == null) {
       return null;
     }
-    final instance = ClientCertResponse(
-      certificatePath: map['certificatePath'],
-    );
+    final instance = ClientCertResponse();
     instance.action = ClientCertResponseAction.fromNativeValue(map['action']);
     instance.androidKeyStoreType = map['keyStoreType'];
     instance.certificatePassword = map['certificatePassword'];
+    if (map['certificatePath'] != null) {
+      instance.certificatePath = map['certificatePath'];
+    }
     instance.keyStoreType = map['keyStoreType'];
+    if (map['selectedCertificate'] != null) {
+      instance.selectedCertificate = map['selectedCertificate'];
+    }
     return instance;
   }
 
@@ -61,6 +82,7 @@ class ClientCertResponse {
       "certificatePassword": certificatePassword,
       "certificatePath": certificatePath,
       "keyStoreType": keyStoreType,
+      "selectedCertificate": selectedCertificate,
     };
   }
 
@@ -71,6 +93,6 @@ class ClientCertResponse {
 
   @override
   String toString() {
-    return 'ClientCertResponse{action: $action, certificatePassword: $certificatePassword, certificatePath: $certificatePath, keyStoreType: $keyStoreType}';
+    return 'ClientCertResponse{action: $action, certificatePassword: $certificatePassword, certificatePath: $certificatePath, keyStoreType: $keyStoreType, selectedCertificate: $selectedCertificate}';
   }
 }